+++ .env.example
@@ -9,6 +9,9 @@ NEXT_PUBLIC_APP_URL='http://localhost:3000'
 NEXTAUTH_URL='http://localhost:3000'
 NEXTAUTH_SECRET='1'
 
+NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY='1'
+CLERK_SECRET_KEY='1'
+
 GITHUB_CLIENT_ID='1'
 GITHUB_CLIENT_SECRET='1'
