+++ apps/nextjs/src/components/github-star.tsx
@@ -8,7 +8,7 @@ export function GitHubStar() {
     <Link href="https://github.com/saasfly/saasfly" target="_blank" rel="Saasfly GitHub">
       <div className="inline-flex items-center gap-1.5 px-3 h-9 border border-input hover:bg-accent hover:text-accent-foreground rounded-full text-sm font-medium">
         <Icons.GitHub className="w-4 h-4"/>
-        <span>2.2K</span>
+        <span>2.4K</span>
       </div>
     </Link>
   )
