+++ apps/nextjs/src/components/modal-provider.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { SignInModal } from "~/components/sign-in-modal";
+import { SignInClerkModal } from "~/components/sign-in-modal-clerk";
 import { useMounted } from "~/hooks/use-mounted";
 
 export const ModalProvider = ({ dict }: { dict: Record<string, string> }) => {
@@ -12,7 +13,8 @@ export const ModalProvider = ({ dict }: { dict: Record<string, string> }) => {
 
   return (
     <>
-      <SignInModal dict={dict} />
+      {/* <SignInModal dict={dict} /> */}
+      <SignInClerkModal dict={dict} />
     </>
   );
 };
