+++ apps/nextjs/src/types/globals.d.ts
@@ -1,0 +1,13 @@
+export {}
+
+declare global {
+  interface CustomJwtSessionClaims {
+    user?: {
+      id: string;
+      name: string;
+      email: string;
+      image: string;
+      isAdmin?: boolean;
+    }
+  }
+}
