+++ package.json
@@ -34,5 +34,8 @@
     "packages/*",
     "tooling/*"
   ],
-  "packageManager": "bun@v1.1.10"
+  "packageManager": "bun@v1.1.10",
+  "dependencies": {
+    "@clerk/nextjs": "^6.19.5"
+  }
 }
