+++ packages/api/src/trpc.ts
@@ -1,12 +1,13 @@
 import type { NextRequest } from "next/server";
 import { initTRPC } from "@trpc/server";
-import { getToken, type JWT } from "next-auth/jwt";
+import { auth } from "@clerk/nextjs/server";
 import { ZodError } from "zod";
 
 import { transformer } from "./transformer";
 
 interface CreateContextOptions {
   req?: NextRequest;
+  auth?: any;
 }
 
 export const createInnerTRPCContext = (opts: CreateContextOptions) => {
@@ -15,9 +16,11 @@ export const createInnerTRPCContext = (opts: CreateContextOptions) => {
   };
 };
 
-export const createTRPCContext = (opts: { req: NextRequest }) => {
+// see: https://clerk.com/docs/references/nextjs/trpc
+export const createTRPCContext = async (opts: { req: NextRequest }) => {
   return createInnerTRPCContext({
     req: opts.req,
+    auth: await auth(),
   });
 };
 
@@ -40,13 +43,7 @@ export const procedure = t.procedure;
 export const mergeRouters = t.mergeRouters;
 
 export const protectedProcedure = procedure.use(async (opts) => {
-  const { req } = opts.ctx;
-  const nreq = req!;
-  const jwt = await handler(nreq);
-  return opts.next({ ctx: { req, userId: jwt?.id } });
+  const { req, auth } = opts.ctx;
+  const { userId } = auth;
+  return opts.next({ ctx: { req, userId: userId } });
 });
-
-async function handler(req: NextRequest): Promise<JWT | null> {
-  // if using `NEXTAUTH_SECRET` env variable, we detect it, and you won't actually need to `secret`
-  return await getToken({ req });
-}
