+++ packages/db/prisma/types.ts
@@ -1,12 +1,11 @@
 import type { ColumnType } from "kysely";
-
-import type { Status, SubscriptionPlan } from "./enums";
-
 export type Generated<T> =
   T extends ColumnType<infer S, infer I, infer U>
     ? ColumnType<S, I | undefined, U>
     : ColumnType<T, T | undefined, T>;
 export type Timestamp = ColumnType<Date, Date | string, Date | string>;
+
+import type { SubscriptionPlan, Status } from "./enums";
 
 export type Account = {
   id: Generated<string>;
